ape: <stdio.h> no longer drags errno, unistd, fcntl and pthread in
<stdio.h> has to see struct _IO_FILE, because FILE is a typedef of it.
It was getting that by including <stdio_impl.h>, musl's *internal*
header, which in this tree opens with

    #include <stdint.h>   #include <stddef.h>  #include <errno.h>
    #include <unistd.h>   #include <fcntl.h>   #include <sys/types.h>
    #include <pthread.h>

because the implementation genuinely wants all of it. So every
translation unit that included <stdio.h> -- nearly all of them -- also
got every O_ and F_ macro, all of unistd and all of pthread.

That is not only untidy. Portable code guards its fallbacks on whether a
name exists, and a name arriving earlier or later than the author
expected changes the answer. libzip's compat.h is the case that prompted
this: it defines O_CLOEXEC as 0 for Windows' sake, having included
nothing that leads to <fcntl.h> -- but <zip.h> asks for stdio a little
later, fcntl arrives through the back door, and the two disagree.

musl's own arrangement restored: struct _IO_FILE moves to a new
<_iofile.h> with only the three headers it needs -- stddef for size_t,
sys/types for off_t, stdint for the lock's intptr_t; struct
__locale_struct stays an incomplete type behind a pointer. <stdio.h>
includes that instead. stdio_impl.h includes it too and keeps its heavy
includes, so the 64 files in ap/stdio that use it are unaffected.

Checked before, not after: every .c in libap that includes <stdio.h>
without <stdio_impl.h> was scanned for uses of the four leaked headers,
counting what lib.h supplies transitively. Nothing in libap depended on
the leak -- the first pass flagged eight files and all eight were the
regex's fault, EOF and ERR matching an errno-constant pattern, Plan 9's
own open() from sys9.h matching POSIX open(), and read.c's O_NONBLOCK
coming from lib.h's <fcntl.h> rather than from stdio.

That is libap. Ported packages are another matter: any that used
open(), errno or pthread while including only <stdio.h> compiled by
accident until now and will want the right #include adding. Those are
real bugs and each is a one-line fix, but a full rebuild is the way to
find them.
